// Frontend/Modulo - Facturación Equipo y Accesorios/validacion-facturas/src/app/models/InvoiceEquipmentAccesories.ts
export interface InvoiceEquipmentAccesories {
    id?: number;
    invoiceNo: string;
    invoiceType: string;
    invoiceStatus: string;
    invoiceStatusName: string;
    billingAccount: string;
    subscriber: string;
    custcode: string;
    emailaddr: string;
    company: string;
    cai: string;
    customerName: string;
    customerType: string;
    documentNo: string;
    diplomaticCardNo: string;
    correlativeOrdenExemptNo: string;
    correlativeCertificateExoNo: string;
    exonerationStatusName: string;
    xml: string;
    path: string;
    address: string;
    warehouse: string;
    agency: string;
    transactionUser: string;
    chargeLocal: string;
    chargeUsd: string;
    exchangeRate: string;
    tax: string;
    discount: string;
    created: string;
    createdBy: string;
}